Training Assessment For Matthew Schmid
I was unable to satisfactory complete the assessment, although I had a working Java filter for interacting with the SQL database, I was unable to display the databases over the web interface. I still need more practice with the html and jsp file types, I am still trying to complete the assessment in my freetime.